add library for managing rings
Add support for managing the rings using Python instead of forking to
call the swift-ring-builder binaries. This provides much more
flexibility, including the ability to define the id of the device.

NOTE: This is a breaking change.

The format of both the ring and the devices have changed from being
lists to dicts.

Instead of having to specify the id of a ring or device with `id` key,
the id is taken from the name of the ring while for devices it is the
key (int) in the dict.

For example, 'object' and `object-0` will be treated as `object.builder`
while `object-1` will be treated as object-1.builder. The id is
automatically gleaned from the `-<int>` suffix.

There are a number of required keys for rings, they include:
 - name
 - part_power
 - replicas

Optional keys include:
 - region (swift library defaults this to '1')
 - zone (swift library defaults this to '1')

For example, rings can be defined in the inventory like so:

    swift_rings:
      account:
        part_power: 10
        region: 1
        replicas: 3
        zone: 1
      container:
        part_power: 10
        replicas: 3
      object:
        part_power: 17
        replicas: 3
      object-1:
        part_power: 17
        replicas: 7
        policy_type: erasure_coding
        ec_num_data_fragments: 4
        ec_num_parity_fragments: 3
        ec_duplication_factor: 1

For devices, these are now specified as a dict where the id of the disk
is the key, which is also gleaned automatically. This ensures that each
disk in the inventory is unique. Each device supports a dict of rings,
defining which rings the device should be added to.

There are a number of required keys for devices in rings, they include:
  - device
  - ip
  - port
  - region
  - weight
  - zone

There are also optional keys, which include:
  - meta
  - replication_ip
  - replication_port

You can specify all of the required keys at the device level, that way
they will be used as defaults, able to be overridden at the device's
ring level.

For example, devices can be defined in the inventory like so:

    swift_devs:
      0:
        device: /dev/sdb
        ip: 127.0.0.1
        port: 8080
        region: 1
        zone: 1
        rings:
          object:
            weight: 100
          object_1:
            weight: 50
      1:
        device: /dev/sdc
        port: 8080
        ip: 127.0.0.2
        rings:
          account:
            region: 1
            weight: 50
            zone: 1
          container:
            weight: 100