Add support for CA cert
This adds support to install and trust the CA Certificate on the hosts,
which should mean that swift commands, as well as dispersion reports,
from the nodes themselves will no-longer fail.

The default variables now include the example ca cert and ca key.
Certificates and keys have also been updated, with CN of swift and SAN
of swift and default IP addresses.