use swift_admin group instead of swift-admin host
Originally the role expected a host called 'swift-admin' this has been
replaced with the first node in the 'swift_admin' group instead.

This means if you don't have to have a machine called 'swift-admin' in
your inventory and you can also re-use one of the Swift nodes if you
prefer.