fix add ssh key task
The SSH key task was set to run_once so that all nodes get it, but also
set to only run when the host was in the admin group.

The chosen host might not be in the admin group and therefore the task
is skipped, causing further steps to fail as they won't have the SSH
key.

This change instead lets it run_once but on any host and fetches the
variable directly from the first host in the admin group.