support more complex device paths
Some disks might not show up under /dev/, for example, LVM or iSCSI
devices.

This patch adds the ability to express the device using its full path,
for example `/dev/sdb` instead of just `sdb`.

The device ring still uses the last part of the device (splitting on '/
') but stores the full device path in the meta field.

When performing disk operations such as formatting and mounting, the
full path is used.