get last 12 chars for disk label
XFS only supports a maximum of 12 chars for a label. This is currently
derived from the last part of the disk path, e.g. `sdb` for `/dev/sdb`.
However, for disks which are more complex, such as mpath, the device may
be too long.

In the patch that added support for complex device names, the label was
set by getting the first 12 chars. However, it's probably better to get
the last 12 chars as that's where integers are more likely to be and so
we're less likely to have conflicts. For labels less than 12 chars, this
makes no difference.

Ultimately, the label should be set to the id of the disk in the Swift
rings, but that requires the library to be finished and so until then,
this patch should help.