add support for firewall
Some machines may have firewalld enabled, so we should support them,
else the cluster appears broken.

This adds new tasks for configuring firewalld and enabling the relevant
ports. This is done on all hosts, regardless of role, however it could
be tightened up if necessary.

There is a new variable that's checked that determines whether the
firewall should be configured. It is set to true by default, meaning the
firewall will be configured, set to false to skip firewall
configuration.

  swift_firewall: true

This assumes the public zone, if this is not the zone in use, then set
the zone with:

  swift_firewall_zone: other

Fixes #9