flush handlers after playbook sections complete
If a playbook run fails, some services can become broken. This is
because while handlers are notified after certain tasks (like writing a
config file) they aren't run until the end of the play and if the
playbook fails, then the handlers aren't run.

Then next time the playbook is successfully run, the config files are not
changed and so the hanlders are not notified.

This patch simply flushes the handlers after each section of the play,
however a proper fix needs to be implemented making the configurations
themselves more robust (with blocks and rescue).

Fixes #10