determine device with outward IP for keepalived
The keepalived config was hard-coded to use `eth0` which meant that it
did not start properly on nodes where the device had a different name.

This change updates the template to loop through all network interfaces
and find the one that matches the outward IP.

If an interface exists that does not have an IP address assigned, the
jinja2 expression will default to localhost so as to avoid a failure.